Refactor muliline braces shared examples to avoid ifs
