+++ spec/rubocop/cop/layout/multiline_hash_brace_layout_spec.rb
@@ -35,6 +35,15 @@ RSpec.describe RuboCop::Cop::Layout::MultilineHashBraceLayout, :config do
     end
   end
 
+  include_examples 'multiline literal brace layout method argument' do
+    let(:open) { '{' }
+    let(:close) { '}' }
+    let(:a) { 'a: 1' }
+    let(:b) { 'b: 2' }
+    let(:multi_prefix) { 'b: ' }
+    let(:multi) { ['[', '1', ']'] }
+  end
+
   include_examples 'multiline literal brace layout trailing comma' do
     let(:open) { '{' }
     let(:close) { '}' }
